Advent of Code 2023 day 4 part 2
This one was fraught with peril, tricky to figure out the proper math when I get stuck and the only other Odin attempt at this problem comes back with an algorithm that doesn't work.  In the end, taking a step back and looking at the basics was what saved me!